fix logic in remove book
