+++ assets/scripts/index.js
@@ -1,2 +1,2 @@
-export { list, addNew, contact } from './navigation.js';
-export { addBook, remove } from './book_funtionalities.js';
+export { addBook, remove } from './modules/book_funtionalities.js';
+export { list, addNew, contact } from './modules/navigation.js';
