+++ assets/scripts/modules/storage.js
@@ -1,4 +1,2 @@
 const data = localStorage.getItem('library');
-let library = JSON.parse(data) || [];
-
-export default library;
+export default data;
