+++ index.html
@@ -60,8 +60,7 @@
     <footer>
         <p class="Copyright">copyright</p>
     </footer>
-    <script src="./assets/scripts/bookclasses.js"></script>
-    <script src="./assets/scripts/navigation.js"></script>
+    <script type="module" src="./assets/scripts/index.js"></script>
 </body>
 
 </html>
